backend/app/schemas/auth.py
from pydantic import BaseModel
from typing import Optional


class Token(BaseModel):
    """Token schema"""
    access_token: str
    token_type: str = "bearer"


class TokenData(BaseModel):
    """Token data schema"""
    username: Optional[str] = None
