backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
minio==7.2.0
aiofiles==23.2.0
python-magic==0.4.27
pydantic-settings==2.0.3
email-validator==2.1.0